implement "storage digests"
A storage digest is the hash of the content digest--the former is used as the
on-disk filename, rather than the content digest. With a 1-way hash function
like sha256, this makes the content digest unrecoverable from persistent
storage.

Provided content digests are not persisted in any way, the effect is it is not
possible to read plaintext the plaintext of files encrypted with the content
digest from storage out of band, unless you already know the content digest of
that plaintext.